requirements.txt
# Core ML dependencies
transformers>=4.36.0
peft>=0.7.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
trl>=0.7.6
datasets>=2.14.7
# llm-lingua>=0.2.1
llmlingua
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
hf_xet

# Text processing
spacy>=3.7.0
sentencepiece>=0.1.99
tokenizers>=0.15.0
jsonlines>=4.0.0
nltk>=3.8.1

# Data science utilities
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.66.0
scipy>=1.11.0

# Experiment tracking
wandb>=0.15.12
tensorboard>=2.14.0
mlflow>=2.8.0

# Development tools
pytest>=7.4.0
black>=23.9.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.1
pytest-cov>=4.1.0

# Documentation
mkdocs>=1.5.3
mkdocs-material>=9.4.6
mkdocstrings>=0.24.0
mkdocstrings-python>=1.7.0

# Optional: for interactive development
ipython>=8.12.0
jupyter>=1.0.0
notebook>=7.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Retriever 
rank-bm25

# Reader
google-genai 
python-dotenv

# Backend
fastapi
uvicorn

# Evaluation
tabulate
hf_xet
rouge_score
llmlingua
# Pre-download spacy model
# Run after installing requirements:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
# python -m spacy download en_core_web_sm